SDK-8780 PR review changes
