+++ src/auth0_server_python/error/__init__.py
@@ -229,6 +229,24 @@ class CustomTokenExchangeErrorCode:
     INVALID_RESPONSE = "invalid_response"
 
 
+class PasskeyError(Auth0Error):
+    """
+    Error raised during passkey authentication operations.
+    """
+    def __init__(self, code: str, message: str, cause=None):
+        super().__init__(message)
+        self.code = code
+        self.name = "PasskeyError"
+        self.cause = cause
+
+
+class PasskeyErrorCode:
+    """Error codes for passkey operations."""
+    CHALLENGE_FAILED = "passkey_challenge_error"
+    TOKEN_EXCHANGE_FAILED = "passkey_token_error"
+    INVALID_RESPONSE = "invalid_response"
+
+
 # =============================================================================
 # MFA Error Classes
 # =============================================================================
