+++ src/auth0_server_python/tests/test_server_client.py
@@ -4816,3 +4816,797 @@ async def test_get_access_token_mfa_required_with_enroll_requirements(mocker):
         assert exc.value.mfa_requirements is not None
     finally:
         ServerClient._fetch_oidc_metadata = original_fetch
+
+
+# =============================================================================
+# PASSKEY AUTHENTICATION
+# =============================================================================
+
+_PASSKEY_SIGNUP_CHALLENGE_RESPONSE = {
+    "auth_session": "session_abc123",
+    "authn_params_public_key": {
+        "challenge": "dGVzdC1jaGFsbGVuZ2U",
+        "rp": {"id": "auth0.local", "name": "Test App"},
+        "user": {"id": "dXNlcl8x", "name": "user@example.com", "displayName": "Jane"},
+        "pubKeyCredParams": [{"type": "public-key", "alg": -7}],
+        "authenticatorSelection": {
+            "residentKey": "required",
+            "userVerification": "preferred",
+        },
+        "timeout": 60000,
+    },
+}
+
+_PASSKEY_LOGIN_CHALLENGE_RESPONSE = {
+    "auth_session": "session_login_xyz",
+    "authn_params_public_key": {
+        "challenge": "bG9naW4tY2hhbGxlbmdl",
+        "rpId": "auth0.local",
+        "timeout": 60000,
+        "userVerification": "preferred",
+    },
+}
+
+_PASSKEY_TOKEN_RESPONSE = {
+    "access_token": "at_passkey_123",
+    "id_token": "eyJ.test.jwt",
+    "token_type": "Bearer",
+    "expires_in": 86400,
+    "scope": "openid profile",
+}
+
+
+def _make_passkey_authn_response():
+    from auth0_server_python.auth_types import PasskeyAuthResponse
+    return PasskeyAuthResponse(
+        id="cred_abc123",
+        raw_id="Y3JlZF9hYmMxMjM",
+        type="public-key",
+        authenticator_attachment="platform",
+        response={
+            "clientDataJSON": "eyJ0eXBlIjoid2ViYXV0aG4uZ2V0In0",
+            "authenticatorData": "SZYN5YgOjGh0NBcPZHZgW4_krrmihjLHmVzzuoMdl2M",
+            "signature": "MEUCIQC",
+            "userHandle": "dXNlcl8x",
+        },
+    )
+
+
+@pytest.mark.asyncio
+async def test_passkey_signup_challenge_success(mocker):
+    from auth0_server_python.auth_types import PasskeySignupChallengeResponse, PasskeyUserProfile
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 200
+    mock_response.json = MagicMock(return_value=_PASSKEY_SIGNUP_CHALLENGE_RESPONSE)
+    mock_post.return_value = mock_response
+
+    result = await client.passkey_signup_challenge(
+        user_profile=PasskeyUserProfile(email="user@example.com", name="Jane Doe"),
+        connection="Username-Password-Authentication",
+    )
+
+    assert isinstance(result, PasskeySignupChallengeResponse)
+    assert result.auth_session == "session_abc123"
+    assert result.authn_params_public_key.challenge == "dGVzdC1jaGFsbGVuZ2U"
+    assert result.authn_params_public_key.rp.id == "auth0.local"
+    assert result.authn_params_public_key.user.display_name == "Jane"
+    assert result.authn_params_public_key.pub_key_cred_params[0].alg == -7
+    assert result.authn_params_public_key.authenticator_selection.resident_key == "required"
+
+    mock_post.assert_awaited_once()
+    args, kwargs = mock_post.call_args
+    assert "/passkey/register" in args[0]
+    body = kwargs["json"]
+    assert body["client_id"] == "test_client_id"
+    assert body["client_secret"] == "test_client_secret"
+    assert body["user_profile"]["email"] == "user@example.com"
+    assert body["user_profile"]["name"] == "Jane Doe"
+    assert body["realm"] == "Username-Password-Authentication"
+
+
+@pytest.mark.asyncio
+async def test_passkey_signup_challenge_user_profile_fields(mocker):
+    from auth0_server_python.auth_types import PasskeyUserProfile
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 200
+    mock_response.json = MagicMock(return_value=_PASSKEY_SIGNUP_CHALLENGE_RESPONSE)
+    mock_post.return_value = mock_response
+
+    await client.passkey_signup_challenge(
+        user_profile=PasskeyUserProfile(
+            email="u@e.com",
+            username="jdoe",
+            phone_number="+1234567890",
+            given_name="Jane",
+            family_name="Doe",
+            nickname="jd",
+            picture="https://example.com/pic.jpg",
+            user_metadata={"role": "admin"},
+        ),
+        organization="org_123",
+    )
+
+    args, kwargs = mock_post.call_args
+    body = kwargs["json"]
+    assert body["user_profile"]["email"] == "u@e.com"
+    assert body["user_profile"]["username"] == "jdoe"
+    assert body["user_profile"]["phone_number"] == "+1234567890"
+    assert body["user_profile"]["given_name"] == "Jane"
+    assert body["user_profile"]["family_name"] == "Doe"
+    assert body["user_profile"]["nickname"] == "jd"
+    assert body["user_profile"]["picture"] == "https://example.com/pic.jpg"
+    assert body["user_profile"]["user_metadata"] == {"role": "admin"}
+    assert body["organization"] == "org_123"
+
+
+@pytest.mark.asyncio
+async def test_passkey_signup_challenge_minimal_body(mocker):
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 200
+    mock_response.json = MagicMock(return_value=_PASSKEY_SIGNUP_CHALLENGE_RESPONSE)
+    mock_post.return_value = mock_response
+
+    await client.passkey_signup_challenge()
+
+    args, kwargs = mock_post.call_args
+    body = kwargs["json"]
+    assert body == {"client_id": "test_client_id", "client_secret": "test_client_secret"}
+    assert "user_profile" not in body
+    assert "realm" not in body
+    assert "organization" not in body
+
+
+@pytest.mark.asyncio
+async def test_passkey_signup_challenge_api_error(mocker):
+    from auth0_server_python.auth_types import PasskeyUserProfile
+    from auth0_server_python.error import PasskeyError
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 403
+    mock_response.json = MagicMock(return_value={
+        "error": "access_denied",
+        "error_description": "Passkey not enabled",
+    })
+    mock_post.return_value = mock_response
+
+    with pytest.raises(PasskeyError) as exc:
+        await client.passkey_signup_challenge(
+            user_profile=PasskeyUserProfile(email="test@example.com")
+        )
+    assert "access_denied" in str(exc.value) or "Passkey not enabled" in str(exc.value)
+
+
+@pytest.mark.asyncio
+async def test_passkey_signup_challenge_non_json_error(mocker):
+    from auth0_server_python.error import PasskeyError
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 502
+    mock_response.json = MagicMock(side_effect=json.JSONDecodeError("bad", "", 0))
+    mock_post.return_value = mock_response
+
+    with pytest.raises(PasskeyError) as exc:
+        await client.passkey_signup_challenge()
+    assert "502" in str(exc.value) or "passkey_challenge_error" in str(exc.value)
+
+
+@pytest.mark.asyncio
+async def test_passkey_signup_challenge_network_error(mocker):
+    from auth0_server_python.error import PasskeyError
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_post.side_effect = Exception("Connection refused")
+
+    with pytest.raises(PasskeyError) as exc:
+        await client.passkey_signup_challenge()
+    assert "Passkey signup challenge failed" in str(exc.value)
+
+
+@pytest.mark.asyncio
+async def test_passkey_login_challenge_success(mocker):
+    from auth0_server_python.auth_types import PasskeyLoginChallengeResponse
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 200
+    mock_response.json = MagicMock(return_value=_PASSKEY_LOGIN_CHALLENGE_RESPONSE)
+    mock_post.return_value = mock_response
+
+    result = await client.passkey_login_challenge(
+        connection="Username-Password-Authentication",
+        organization="org_abc",
+    )
+
+    assert isinstance(result, PasskeyLoginChallengeResponse)
+    assert result.auth_session == "session_login_xyz"
+    assert result.authn_params_public_key.challenge == "bG9naW4tY2hhbGxlbmdl"
+    assert result.authn_params_public_key.rp_id == "auth0.local"
+    assert result.authn_params_public_key.user_verification == "preferred"
+
+    args, kwargs = mock_post.call_args
+    body = kwargs["json"]
+    assert body["client_id"] == "test_client_id"
+    assert body["realm"] == "Username-Password-Authentication"
+    assert body["organization"] == "org_abc"
+
+
+@pytest.mark.asyncio
+async def test_passkey_login_challenge_with_username(mocker):
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 200
+    mock_response.json = MagicMock(return_value=_PASSKEY_LOGIN_CHALLENGE_RESPONSE)
+    mock_post.return_value = mock_response
+
+    await client.passkey_login_challenge(username="jane@example.com")
+
+    args, kwargs = mock_post.call_args
+    body = kwargs["json"]
+    assert body["username"] == "jane@example.com"
+
+
+@pytest.mark.asyncio
+async def test_passkey_login_challenge_api_error(mocker):
+    from auth0_server_python.error import PasskeyError
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 400
+    mock_response.json = MagicMock(return_value={
+        "error": "invalid_request",
+        "error_description": "Missing client_id",
+    })
+    mock_post.return_value = mock_response
+
+    with pytest.raises(PasskeyError):
+        await client.passkey_login_challenge()
+
+
+@pytest.mark.asyncio
+async def test_passkey_login_challenge_network_error(mocker):
+    from auth0_server_python.error import PasskeyError
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_post.side_effect = Exception("timeout")
+
+    with pytest.raises(PasskeyError):
+        await client.passkey_login_challenge()
+
+
+@pytest.mark.asyncio
+async def test_signin_with_passkey_success(mocker):
+    from auth0_server_python.auth_types import PasskeyTokenResponse
+    from auth0_server_python.error import PasskeyError
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mocker.patch.object(
+        client,
+        "_get_oidc_metadata_cached",
+        return_value={"token_endpoint": "https://auth0.local/oauth/token"},
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 200
+    mock_response.json = MagicMock(return_value=_PASSKEY_TOKEN_RESPONSE)
+    mock_post.return_value = mock_response
+    authn_response = _make_passkey_authn_response()
+
+    result = await client.signin_with_passkey(
+        auth_session="session_xyz",
+        authn_response=authn_response,
+        scope="openid profile",
+        audience="https://api.example.com",
+        connection="Username-Password-Authentication",
+        organization="org_abc",
+    )
+
+    assert isinstance(result, PasskeyTokenResponse)
+    assert result.access_token == "at_passkey_123"
+    assert result.token_type == "Bearer"
+    assert abs(result.expires_at - (int(time.time()) + 86400)) <= 2
+
+    mock_post.assert_awaited_once()
+    args, kwargs = mock_post.call_args
+    body = kwargs["json"]
+    assert body["grant_type"] == "urn:okta:params:oauth:grant-type:webauthn"
+    assert body["client_id"] == "test_client_id"
+    assert body["client_secret"] == "test_client_secret"
+    assert body["auth_session"] == "session_xyz"
+    assert body["scope"] == "openid profile"
+    assert body["audience"] == "https://api.example.com"
+    assert body["realm"] == "Username-Password-Authentication"
+    assert body["organization"] == "org_abc"
+    assert body["authn_response"]["rawId"] == "Y3JlZF9hYmMxMjM"
+    assert body["authn_response"]["authenticatorAttachment"] == "platform"
+    assert "raw_id" not in body["authn_response"]
+
+
+@pytest.mark.asyncio
+async def test_signin_with_passkey_uses_json_content_type(mocker):
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mocker.patch.object(
+        client,
+        "_get_oidc_metadata_cached",
+        return_value={"token_endpoint": "https://auth0.local/oauth/token"},
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 200
+    mock_response.json = MagicMock(return_value=_PASSKEY_TOKEN_RESPONSE)
+    mock_post.return_value = mock_response
+
+    await client.signin_with_passkey(
+        auth_session="s",
+        authn_response=_make_passkey_authn_response(),
+    )
+
+    args, kwargs = mock_post.call_args
+    assert "json" in kwargs
+    assert "data" not in kwargs
+
+
+@pytest.mark.asyncio
+@pytest.mark.parametrize("auth_session", [None, ""])
+async def test_signin_with_passkey_missing_auth_session(auth_session):
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    with pytest.raises(MissingRequiredArgumentError):
+        await client.signin_with_passkey(
+            auth_session=auth_session,
+            authn_response=_make_passkey_authn_response(),
+        )
+
+
+@pytest.mark.asyncio
+async def test_signin_with_passkey_missing_authn_response():
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    with pytest.raises(MissingRequiredArgumentError):
+        await client.signin_with_passkey(
+            auth_session="session_abc",
+            authn_response=None,
+        )
+
+
+@pytest.mark.asyncio
+async def test_signin_with_passkey_api_error(mocker):
+    from auth0_server_python.error import PasskeyError
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mocker.patch.object(
+        client,
+        "_get_oidc_metadata_cached",
+        return_value={"token_endpoint": "https://auth0.local/oauth/token"},
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 401
+    mock_response.json = MagicMock(return_value={
+        "error": "invalid_grant",
+        "error_description": "Invalid auth_session",
+    })
+    mock_post.return_value = mock_response
+
+    with pytest.raises(PasskeyError) as exc:
+        await client.signin_with_passkey(
+            auth_session="expired_session",
+            authn_response=_make_passkey_authn_response(),
+        )
+    assert "invalid_grant" in str(exc.value) or "Invalid auth_session" in str(exc.value)
+
+
+@pytest.mark.asyncio
+async def test_signin_with_passkey_missing_token_endpoint(mocker):
+    from auth0_server_python.error import PasskeyError
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mocker.patch.object(client, "_get_oidc_metadata_cached", return_value={})
+
+    with pytest.raises(PasskeyError) as exc:
+        await client.signin_with_passkey(
+            auth_session="session",
+            authn_response=_make_passkey_authn_response(),
+        )
+    assert "token endpoint" in str(exc.value).lower()
+
+
+@pytest.mark.asyncio
+async def test_signin_with_passkey_network_error(mocker):
+    from auth0_server_python.error import PasskeyError
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mocker.patch.object(
+        client,
+        "_get_oidc_metadata_cached",
+        return_value={"token_endpoint": "https://auth0.local/oauth/token"},
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_post.side_effect = Exception("Connection reset")
+
+    with pytest.raises(PasskeyError):
+        await client.signin_with_passkey(
+            auth_session="session",
+            authn_response=_make_passkey_authn_response(),
+        )
+
+
+@pytest.mark.asyncio
+async def test_signin_with_passkey_no_client_secret(mocker):
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="public_client",
+        client_secret=None,
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret",
+    )
+    mocker.patch.object(
+        client,
+        "_get_oidc_metadata_cached",
+        return_value={"token_endpoint": "https://auth0.local/oauth/token"},
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 200
+    mock_response.json = MagicMock(return_value=_PASSKEY_TOKEN_RESPONSE)
+    mock_post.return_value = mock_response
+
+    from auth0_server_python.auth_types import PasskeyAuthResponse
+    authn_resp = PasskeyAuthResponse(
+        id="cred",
+        raw_id="cmF3",
+        type="public-key",
+        response={"clientDataJSON": "abc", "authenticatorData": "def", "signature": "ghi"},
+    )
+    await client.signin_with_passkey(auth_session="session", authn_response=authn_resp)
+
+    args, kwargs = mock_post.call_args
+    body = kwargs["json"]
+    assert "client_secret" not in body
+    assert body["client_id"] == "public_client"
+
+
+def test_passkey_signup_challenge_repr_redacts_auth_session():
+    from auth0_server_python.auth_types import PasskeySignupChallengeResponse
+    resp = PasskeySignupChallengeResponse.model_validate(_PASSKEY_SIGNUP_CHALLENGE_RESPONSE)
+    repr_str = repr(resp)
+    assert "session_abc123" not in repr_str
+    assert "[REDACTED]" in repr_str
+
+
+def test_passkey_login_challenge_repr_redacts_auth_session():
+    from auth0_server_python.auth_types import PasskeyLoginChallengeResponse
+    resp = PasskeyLoginChallengeResponse.model_validate(_PASSKEY_LOGIN_CHALLENGE_RESPONSE)
+    repr_str = repr(resp)
+    assert "session_login_xyz" not in repr_str
+    assert "[REDACTED]" in repr_str
+
+
+def test_passkey_token_response_repr_redacts_tokens():
+    from auth0_server_python.auth_types import PasskeyTokenResponse
+    resp = PasskeyTokenResponse(
+        access_token="secret_at_value",
+        token_type="Bearer",
+        expires_in=86400,
+        id_token="secret_id_token",
+        refresh_token="secret_rt_value",
+    )
+    repr_str = repr(resp)
+    assert "secret_at_value" not in repr_str
+    assert "secret_id_token" not in repr_str
+    assert "secret_rt_value" not in repr_str
+    assert "[REDACTED]" in repr_str
+    assert "86400" in repr_str
+
+
+@pytest.mark.asyncio
+async def test_signin_with_passkey_preserves_server_expires_at(mocker):
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mocker.patch.object(
+        client,
+        "_get_oidc_metadata_cached",
+        return_value={"token_endpoint": "https://auth0.local/oauth/token"},
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 200
+    mock_response.json = MagicMock(return_value={
+        "access_token": "at_123",
+        "token_type": "Bearer",
+        "expires_in": 3600,
+        "expires_at": 9999999999,
+    })
+    mock_post.return_value = mock_response
+
+    from auth0_server_python.auth_types import PasskeyTokenResponse
+    result = await client.signin_with_passkey(
+        auth_session="session",
+        authn_response=_make_passkey_authn_response(),
+    )
+    assert result.expires_at == 9999999999
+
+
+@pytest.mark.asyncio
+async def test_signin_with_passkey_missing_expires_at_calculates(mocker):
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mocker.patch.object(
+        client,
+        "_get_oidc_metadata_cached",
+        return_value={"token_endpoint": "https://auth0.local/oauth/token"},
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 200
+    mock_response.json = MagicMock(return_value={
+        "access_token": "at_123",
+        "token_type": "Bearer",
+        "expires_in": 60,
+    })
+    mock_post.return_value = mock_response
+
+    result = await client.signin_with_passkey(
+        auth_session="session",
+        authn_response=_make_passkey_authn_response(),
+    )
+    assert abs(result.expires_at - (int(time.time()) + 60)) <= 2
+
+
+@pytest.mark.asyncio
+async def test_signin_with_passkey_dpop_attaches_proof_header(mocker):
+    import base64
+    import json as _json
+    from jwcrypto import jwk as jwk_module
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mocker.patch.object(
+        client,
+        "_get_oidc_metadata_cached",
+        return_value={"token_endpoint": "https://auth0.local/oauth/token"},
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 200
+    mock_response.json = MagicMock(return_value=_PASSKEY_TOKEN_RESPONSE)
+    mock_post.return_value = mock_response
+
+    dpop_key = jwk_module.JWK.generate(kty="EC", crv="P-256")
+    await client.signin_with_passkey(
+        auth_session="session_xyz",
+        authn_response=_make_passkey_authn_response(),
+        dpop_key=dpop_key,
+    )
+
+    args, kwargs = mock_post.call_args
+    assert "DPoP" in kwargs["headers"]
+
+    # Decode proof and assert no ath claim (token endpoint proof — RFC 9449 §4.2)
+    proof = kwargs["headers"]["DPoP"]
+    payload_b64 = proof.split(".")[1]
+    padding = 4 - len(payload_b64) % 4
+    payload = _json.loads(base64.urlsafe_b64decode(payload_b64 + "=" * padding))
+    assert "ath" not in payload
+    assert "jti" in payload
+    assert payload["htm"] == "POST"
+    assert payload["htu"] == "https://auth0.local/oauth/token"
+
+
+@pytest.mark.asyncio
+async def test_signin_with_passkey_dpop_nonce_retry(mocker):
+    import base64
+    import json as _json
+    from jwcrypto import jwk as jwk_module
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mocker.patch.object(
+        client,
+        "_get_oidc_metadata_cached",
+        return_value={"token_endpoint": "https://auth0.local/oauth/token"},
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+
+    nonce_response = AsyncMock()
+    nonce_response.status_code = 401
+    nonce_response.headers = {"DPoP-Nonce": "server-nonce-abc"}
+    nonce_response.json = MagicMock(return_value={"error": "use_dpop_nonce"})
+
+    success_response = AsyncMock()
+    success_response.status_code = 200
+    success_response.json = MagicMock(return_value=_PASSKEY_TOKEN_RESPONSE)
+
+    mock_post.side_effect = [nonce_response, success_response]
+
+    dpop_key = jwk_module.JWK.generate(kty="EC", crv="P-256")
+    result = await client.signin_with_passkey(
+        auth_session="session_xyz",
+        authn_response=_make_passkey_authn_response(),
+        dpop_key=dpop_key,
+    )
+
+    assert mock_post.await_count == 2
+    assert result.access_token == "at_passkey_123"
+
+    # Second call must include the nonce in the DPoP proof
+    second_call_kwargs = mock_post.call_args_list[1][1]
+    proof = second_call_kwargs["headers"]["DPoP"]
+    payload_b64 = proof.split(".")[1]
+    padding = 4 - len(payload_b64) % 4
+    payload = _json.loads(base64.urlsafe_b64decode(payload_b64 + "=" * padding))
+    assert payload["nonce"] == "server-nonce-abc"
+
+
+@pytest.mark.asyncio
+async def test_signin_with_passkey_without_dpop_no_dpop_header(mocker):
+    client = ServerClient(
+        domain="auth0.local",
+        client_id="test_client_id",
+        client_secret="test_client_secret",
+        state_store=AsyncMock(),
+        transaction_store=AsyncMock(),
+        secret="test-secret-value",
+    )
+    mocker.patch.object(
+        client,
+        "_get_oidc_metadata_cached",
+        return_value={"token_endpoint": "https://auth0.local/oauth/token"},
+    )
+    mock_post = mocker.patch("httpx.AsyncClient.post", new_callable=AsyncMock)
+    mock_response = AsyncMock()
+    mock_response.status_code = 200
+    mock_response.json = MagicMock(return_value=_PASSKEY_TOKEN_RESPONSE)
+    mock_post.return_value = mock_response
+
+    await client.signin_with_passkey(
+        auth_session="session_xyz",
+        authn_response=_make_passkey_authn_response(),
+    )
+
+    args, kwargs = mock_post.call_args
+    assert "DPoP" not in kwargs.get("headers", {})
